% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{build_book}
\alias{build_book}
\title{Build book from vignettes}
\usage{
build_book(path = "inst/report", path.v = "vignettes",
  output_format = c("bookdown::gitbook", "bookdown::pdf_book"))
}
\arguments{
\item{path}{Path of the book}

\item{path.v}{Path to vignettes folder to copy in the book folder}

\item{output_format}{Output format of the book. "bookdown::gitbook", "bookdown::pdf_book"}
}
\description{
Build book from vignettes
}
