% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pkgdown.R
\name{build_pkgdown}
\alias{build_pkgdown}
\title{Build pkgdown site and move to inst}
\usage{
build_pkgdown(
  move = TRUE,
  clean_before = TRUE,
  clean_after = TRUE,
  yml,
  favicon,
  preview = NA,
  ...
)
}
\arguments{
\item{move}{Logical. Whether to move the "docs" folder in "inst" to be kept in the package}

\item{clean_before}{Logical. Whether to empty the "docs" and "inst/docs" prior to build site}

\item{clean_after}{Logical. Whether to remove the original "docs" folder at the root of the project}

\item{yml}{path to custom "_pkgdown.yml" file}

\item{favicon}{path to favicon}

\item{preview}{If \code{TRUE}, or \code{is.na(preview) && interactive()}, will preview
freshly generated section in browser.}

\item{...}{Other parameters needed by \code{\link{build_site}}}
}
\description{
If "docs" is not in "inst" folder, it will not be available to the users
}
