% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{create_biblio_file}
\alias{create_biblio_file}
\title{Create bibliography md file from a vector of packages}
\usage{
create_biblio_file(packages, out.dir, output = c("packages",
  "references"), to = c("html", "markdown"), custom.md, edit = TRUE)
}
\arguments{
\item{packages}{vector of packages names}

\item{out.dir}{Directory where to save output md file}

\item{output}{Vector with "packages" (list of packages as bullet points),
"references" (list of citation references) or both "packages" and "references"}

\item{to}{Format to convert to. "html" or "markdown"}

\item{custom.md}{Vector of markdown text to add to the document before rendering}

\item{edit}{Logical. Whether to open output md file for manual edition}
}
\description{
Create bibliography md file from a vector of packages
}
\examples{
packages <- c("rmarkdown", "attachment")
create_biblio_file(packages, out.dir = tempdir())
}
