% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{create_book}
\alias{create_book}
\title{Create a directory for a book}
\usage{
create_book(path = "inst/report", clean = FALSE, template)
}
\arguments{
\item{path}{Book directory}

\item{clean}{Logical. Whether to remove Chapter Rmd files.}

\item{template}{A folder with personal template files for a bookdown site}
}
\description{
Create a directory for a book
}
\examples{
\dontrun{
book_folder <- tempdir()
create_book(path = book_folder, clean = TRUE)
}

}
