% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{create_pkg_biblio_file}
\alias{create_pkg_biblio_file}
\title{Create bibliography md file from a package DESCRIPTION file}
\usage{
create_pkg_biblio_file(
  path = "DESCRIPTION",
  out.dir,
  output = c("packages", "references"),
  to = c("html", "markdown"),
  custom.md,
  edit = TRUE
)
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{out.dir}{Directory where to save output md file}

\item{output}{Vector with "packages" (list of packages as bullet points),
"references" (list of citation references) or both "packages" and "references"}

\item{to}{Format to convert to. "html", "markdown" or "raw" text}

\item{custom.md}{Vector of markdown text to add to the document before rendering}

\item{edit}{Logical. Whether to open output md file for manual edition}
}
\description{
Create bibliography md file from a package DESCRIPTION file
}
