% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_description.R
\name{create_pkg_desc_file}
\alias{create_pkg_desc_file}
\title{Create package description file}
\usage{
create_pkg_desc_file(
  path = "DESCRIPTION",
  source = c("archive", "github", "git"),
  url,
  out.dir,
  to = c("html", "markdown", "raw"),
  edit = TRUE
)
}
\arguments{
\item{path}{Path to description file}

\item{source}{Will the package be delivered as archive, from github or from another git?}

\item{url}{url of the repository if not in URL in DESCRIPTION. Used when source is github or git.}

\item{out.dir}{Path to save output file}

\item{to}{Format to convert to. "html", "markdown" or "raw" text}

\item{edit}{Logical. Whether to open file for edition}
}
\description{
Create package description file
}
