% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_readme.R
\name{generate_readme_rmd}
\alias{generate_readme_rmd}
\title{Create and update a custom README.Rmd with installation instructions}
\usage{
generate_readme_rmd(
  base_path = usethis::proj_get(),
  parts = c("description", "installation"),
  source = c("archive", "github", "git"),
  edit = TRUE,
  url
)
}
\arguments{
\item{base_path}{Path of the package to generate Readme from}

\item{parts}{Parts to modify in the Readme: c("description", "installation")}

\item{source}{Will the package be delivered as archive, from github or from another git?}

\item{edit}{Logical. Whether to open README.Rmd}

\item{url}{url of the repository if not in URL in DESCRIPTION. Used when source is github or git.}
}
\value{
A README.Rmd file
}
\description{
Create and update a custom README.Rmd with installation instructions
}
\examples{
\dontrun{
generate_readme_rmd()
}
}
