% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{open_guide_function}
\alias{open_guide_function}
\title{Create a function inside R folder in your package to open userguide}
\usage{
open_guide_function(path = "inst/report")
}
\arguments{
\item{path}{Path to the guide inside your package}
}
\description{
Create a function inside R folder in your package to open userguide
}
