% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pkgdown.R
\name{open_pkgdown_function}
\alias{open_pkgdown_function}
\title{Create a function inside R folder in your package to open pkgdown website}
\usage{
open_pkgdown_function(path = "inst/docs")
}
\arguments{
\item{path}{Path to the pkgdown inside your package}
}
\description{
Create a function inside R folder in your package to open pkgdown website
}
