% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papillon-package.R
\docType{package}
\name{papillon-package}
\alias{papillon}
\alias{papillon-package}
\title{papillon: Build And Highlight Package Documentation With Customized Templates}
\description{
Helps creating company visual identity. Build your templates for your package vignettes, reports with {bookdown}, package presentation with {pkgdown}. Highlight your documentation in your projects. Allow for code folding.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Thinkr-open/papillon}
  \item Report bugs at \url{https://github.com/Thinkr-open/papillon/issues}
}

}
\author{
\strong{Maintainer}: Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

}
\keyword{internal}
