% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{build_book}
\alias{build_book}
\title{Build book from vignettes}
\usage{
build_book(
  path = "inst/report",
  path.v = "vignettes",
  output_format = c("bookdown::gitbook", "bookdown::pdf_document2"),
  clean_before = TRUE,
  clean_after = TRUE,
  keep_rmd = "index\\\\.Rmd$|zzz-references\\\\.Rmd$",
  clean = TRUE
)
}
\arguments{
\item{path}{Path of the book}

\item{path.v}{Path to vignettes folder to copy in the book folder}

\item{output_format}{Output format of the book. "bookdown::gitbook", "bookdown::pdf_document2"}

\item{clean_before}{Logical. Whether to remove all Rmd (except keep_rmd) before build}

\item{clean_after}{Logical. Whether to remove all Rmd (except keep_rmd) after build}

\item{keep_rmd}{Regex to list Rmd filenames to keep if clean_before or clean_after is TRUE. You'd better keep index.Rmd.}

\item{clean}{Whether to delete the possible output files. If \code{FALSE},
simply print out a list of files/directories that should probably be
deleted. You can set the global option \code{bookdown.clean_book = TRUE} to
force this function to delete files. You are recommended to take a look at
the list of files at least once before actually deleting them, i.e. run
\code{clean_book(FALSE)} before \code{clean_book(TRUE)}.}
}
\description{
Build book from vignettes
}
