% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{out_from_lines}
\alias{out_from_lines}
\title{write output in utf8}
\usage{
out_from_lines(out_lines, to = c("html", "markdown"), out.dir,
  filename = "file", edit = TRUE)
}
\arguments{
\item{out_lines}{vector of html lines to write}

\item{to}{Format to convert to. "html" or "markdown"}

\item{out.dir}{Directory where to save output md file}

\item{filename}{filename without extentions}

\item{edit}{Logical. Whether to open file for edition}
}
\description{
write output in utf8
}
